perf(tmdb): add cache TTL to movies, matching the serials pattern
movies-cache.service.ts's findOrCreate returned the cached DB row
forever once created, unlike serials-cache.service.ts which already
checks cachedAt staleness. Movies are far more static than still-airing
series, so this uses a 7-day TTL rather than the series' 6h one.

Extracted the generic TTL-cache-with-in-flight-dedupe logic out of
tmdb-cache.helper.ts into infrastructure/cache/ttl-cache.helper.ts so
it isn't TMDB-specific by name/location; tmdb-cache.helper.ts now just
wraps it with TMDB's existing defaults, so none of its call sites
needed to change.
